[libbeat] Improve syslog parser/processor error handling (#31798)
- Shifted most of the validation logic to the host language instead of
the Ragel parser. This allows for the message to be fully parsed even if
there are validation errors with one or more fields.
- Errors are accumulated as they occur, any fields that have already been
parsed and validated will be returned, alongside error.message.
- Priority field for RFC 3164 messages are now optional, as some
syslog providers will omit this part of the header (especially when writing to files).
- Structured data fields for RFC 5424 messages are processed in a
second pass. If the structured data format doesn't fit the RFC, it will be
prepended to the message, separated by a space.
- Structured data now uses map[string]interface{}
- Update testify to v1.7.1 dependency for assert.ErrorContains support

(cherry picked from commit cabc8badb964b7bc332bac1226540fca42a2f884)

# Conflicts:
#	libbeat/processors/syslog/syslog_test.go
#	libbeat/reader/syslog/message.go
#	libbeat/reader/syslog/message_test.go
#	libbeat/reader/syslog/syslog.go
#	libbeat/reader/syslog/syslog_test.go
#	libbeat/reader/syslog/util_test.go